separate the Datacube to a separate module `datacube`
